Add initial programmer code
This is a python program that generates flash contents, programs it
using dfu-util and registers the device with TTN. There's still some
rough edges, but the basics should work. This only supports the new
MJS2020 board for now, since the original board needs different tooling
for flashing and needs additional hardware for clock calibration.